# Conflicts:
#	HymsonMES/src/Libraries/Hymson.MES.CoreServices/Services/Common/ManuCommon/IManuCommonService.cs
#	HymsonMES/src/Libraries/Hymson.MES.CoreServices/Services/Common/ManuCommon/ManuCommonService.cs
#	HymsonMES/src/Presentation/Hymson.MES.Api/Hymson.MES.Api.csproj
#	HymsonMES/src/Presentation/Hymson.MES.Api/appsettings.Development.json

+++ HymsonMES/src/Libraries/Hymson.MES.Data/Repositories/Manufacture/ManuProductBadRecord/View/ManuProductBadRecordReportView.cs
@@ -14,7 +14,8 @@ namespace Hymson.MES.Data.Repositories.Manufacture
         /// <summary>
         /// 汇总数量
         /// </summary>
-        public decimal Num { get; set; }
+        //public decimal Num { get; set; }
+        public int Num { get; set; }
     }
 
     public class ManuProductBadRecordLogReportView : BaseEntity
